refactor: search agents
